+++ src/app.ts
@@ -14,11 +14,11 @@ server.use(cors());
 server.use(morgan("dev"));
 server.use(express.json());
 
-if (!process.env.MP_ACCESS_TOKEN) {
+if (!process.env.ACCESS_TOKEN_MP) {
   throw new Error('MP_ACCESS_TOKEN is required');
 }
 // Initialize the client object
-const client = new MercadoPagoConfig({ accessToken: process.env.MP_ACCESS_TOKEN });
+const client = new MercadoPagoConfig({ accessToken: process.env.ACCESS_TOKEN_MP });
 
 server.use("/", router);
