+++ src/routes/donation.route.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { donation, subscription } from "../controller/paymentMp";
+import { donation, subscription } from "../controller/mp/payment.controller";
 
 const donationRouter = Router();
